+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - Changed blood spray particles to be less over the top by default.
 - Made blood spatters slightly smaller.
 - Blood spray will go in the opposite direction of the damage (with some random variance).
+- Added a Common Config option for damage type debug messages.
+- Added Client Config options for blood spray friction and blood mist scale.
 
 ## V2.0.2 Datapack Structure Hotfix
 - Fixed an issue with the built-in datapack structure. Changes `entity_type` to `entity_types` to be compliant with